Commit #2
Changes:
Recipe File
On line 25, I added conformance to the Codable protocol.

RecipeDataService File
This is a new file to the project.
On line 12 i created a constant to store the API URL string.

On lines 15-21 we created a function that attempts to return a URL. In the case where we cannot create the URL a custom error is thrown.

On lines 23-41 we call the url creation function. The returned object is a tuple of data and a URLResponse. If the response is valid we attempt to decode the data into a RecipesAPIResponse object.
If the decoding is successful we return the recipes property of the RecipesAPIResponse object.

On lines 43-46 we created a custom enum that conforms to Error for various error situations.

HomeViewModel File
Created a simple view model that attempts to fetch the recipes from the API and then publishes them to a scrollview in the ContentView.

Next Step:
Publish the array of recipes via the main thread.