Commit #4
Added a CacheService protocol and class to manage saving and fetching data from the cache.

in the view model I created a fetchImageData function that takes in the imageURL String as a parameter. I use that string as the key for fetching and storing values in the cache. If we fail to retrieve the data from the cache we proceed to download the data from the API via the RecipeDataService.
Once downloaded it is stored in the cache with the url String as its key.

I created a RecipeTile view that has a number of checks to determine what code should run. The view has an  property of an optional UIImage. If that UIImage has value then an Image that takes in a uiImage as a parameter is created.
The view also has a boolean property that is is set to true if we have tried to fetch the imageData and caught an error. If this is true then a placeholder image is shown.

If neither of those conditions are true then a ProgressView is shown and a task is created to attempt to fetch the imageData. If we are successful in that then the optional uiImage is set to a new UIImage that takes in the returned data as a parameter.