Commit #5
Experimented with different UI layouts I want to use the refreshable modifier so I've gone with a Instagram style of showing the photos in a single list.
Created a RecipeRow view but it is currently not in use and may not be used.

Extracted the logic to download the image data to a function in the view to improve readability.