Commit #7
RecipeDataService
Lines 12-13 I added the URL Strings for the endpoints that return errors and a blank array

Line 16 I added a variable that is of an enum type to assist in testing.

I relocated several functions to the bottom of the file for improved readability.

Lines 71-82 contain a function that switches on the TestingStatus enum and returns a String to be used in the creation of a URL for the fetchAllRecipes function.

Lines 84-95 contain a function that switches on the TestingStatus enum and returns a String to be used in the creation of a URL for the fetchImageData function.

Lines 102-107 contain the TestingStatus enum.

RecipeDataServiceTests

Lines 16-19 confirm that the returned array is not empty

Lines 21-26 confirm that an invalidResponse error is thrown in the case of a bad server response

Lines 28-36 confirm that a invalidURL error is thrown if the URL fails to be created

Lines 38-45 confirm that the getImageData function returns data

Lines 47-55 confirm that a invalidURL error is thrown if the URL fails to be created

HomeViewModelTests

Lines 15-20 confirm that showAlert is true in the event of an error

Lines 22-27 confirm that showAlert is false in the event of a successful API call

Lines 29-34 confirm that the alertMessage is not blank in the event of an error

Lines 36-41 confirm that the alertMessage is blank in the event of a successful API call

Lines 43-48 confirm that the array of recipes is updated in the event of a successful API call.