Commit #9
Constants
Created a Constants file to store the urlStrings of various endpoints for testing purposes

HomeViewModel
Added the private(set) keyword to multiple published properties so that they can be read but not mutated.
Renamed showEmptyView to showEmptyResultsView

Recipe
removed multi-line comment containing JSON data

CacheService
Removed the ObservableObject conformance
Changed the totalOverallCost to 1024 * 1024 * 50

RecipeDataService
Removed the urlStringWithError and urlStringWithEmptyResults properties
Removed the private keyword from the urlString and removed the default value.
Removed the TestingStatus variable.

Improved the check of the response code in both the fetchAllRecipes and getImageData functions.

Marked both functions that create URLs as internal

Deleted the getAllRecipesURLString and getImageURLString functions

Extracted the RecipeDataServiceErrors enum to its own file.

Deleted the TestingStatus enum

RecipeDataServiceErrors
Added localized descriptions of the errors

HomeView
Added the private keyword to the vm declaration
wrapped view in a VStack and applied the task and navigation title modifiers to the VStack
Changed the alert title to the alert message published from the vm

RecipeTile
Added the private keyword to the uiImage and failedToLoad properties.

RecipeDataServiceTests
Removed references to the testing status of the dataService
Initialized a new data service with each test and passed in the appropriate urlString for each test.

HomeViewModelTests
Removed references to the testing status of the dataService
Initialized a new data service with each test and passed in the appropriate urlString for each test.